fastapi/requirements.txt
uvicorn==0.17.6
fastapi==0.75.0
psycopg2-binary
python-dotenv